move unit tests into a new directory and alter package.json scripts to separate unit from integration during CI
