+++ package.json
@@ -19,12 +19,12 @@
     "linOLD": "concurrently -p '·{name}·' -n CODE,COMMITS -c magenta.inverse 'npm:lint:code' 'npm:lint:commits'",
     "lint:code": "eslint --ext .tsx,.ts .",
     "lint:commits": "commitlint -f $(git log --reverse --pretty=%H | head -1) -t $(git log -1 --pretty=%H)",
-    "test": "export $(cat .test.env | xargs) && USE_FAKE_PORTS='true' jest --group=-bq",
-    "testWatch": "export $(cat .test.env | xargs) && USE_FAKE_PORTS='true' jest --watch",
-    "testRealPorts": "export $(cat .test.env | xargs) && npm run dev:docker && jest --runInBand",
-    "testRealPortsReplit": "jest --runInBand",
-    "testWatchRealPorts": "export $(cat .test.env | xargs) && npm run dev:docker && jest --watch --runInBand",
-    "testWatchRealPortsReplit": "jest --watch --runInBand",
+    "test": "export $(cat .test.env | xargs) && USE_FAKE_PORTS='true' jest __tests__/unit/* --group=-bq",
+    "testWatch": "export $(cat .test.env | xargs) && USE_FAKE_PORTS='true' jest __tests__/unit/* --watch",
+    "testRealPorts": "export $(cat .test.env | xargs) && npm run dev:docker && jest __tests__/unit/* --runInBand",
+    "testWatchRealPorts": "export $(cat .test.env | xargs) && npm run dev:docker && jest __tests__/unit/* --watch --runInBand",
+    "testRealPortsReplit": "jest __tests__/unit/* --runInBand",
+    "testWatchRealPortsReplit": "jest __tests__/unit/* --watch --runInBand",
     "testIntegration": "export $(cat .dev.env | xargs) && jest __tests__/integration/* --runInBand",
     "testIntegrationReplit": "jest __tests__/integration/* --runInBand",
     "prepare": "husky install"
